1.extraction of data structures 2.Simple web services
