+++ Cargo.toml
@@ -1,4 +1,6 @@
 [workspace]
 members = [
     "scanner",
+    "explorer",
+    "module",
 ]
