+++ module/Cargo.toml
@@ -1,12 +1,11 @@
 [package]
-name = "explorer"
+name = "module"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-axum = "0.4.3"
-async-graphql = "2.9.9"
-module = { path = "../module" }
-tokio = { version = "1.0", features = ["full"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+chrono = { version = "0.4.19", features = ["serde"] }
