+++ module/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod db;
+pub mod rpc;
