+++ module/src/rpc/mod.rs
@@ -1,0 +1,3 @@
+pub mod block;
+pub mod tx;
+pub mod validator;
