+++ scanner/Cargo.toml
@@ -16,3 +16,4 @@ base64 = "0.13.0"
 clap = { version = "3.0.5", features = ["derive"] }
 sha2 = "0.10.1"
 hex = "0.4.3"
+module = { path = "../module" }
