+++ scanner/src/utils/block.rs
@@ -1,59 +1,18 @@
-use serde::Deserialize;
-
 use crate::Result;
 
 use super::RPCResponse;
 
-#[derive(Deserialize, Debug)]
-pub struct BlockId {
-    pub hash: String,
-}
+use module::rpc::block::BlockRPC as module_block_rpc;
 
-#[derive(Deserialize, Debug)]
-pub struct BlockHeader {
-    pub height: String,
-    pub time: String,
-    pub app_hash: String,
-    pub proposer_address: String,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct Signature {
-    pub validator_address: Option<String>,
-    pub timestamp: Option<String>,
-    pub signature: Option<String>,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct LastCommit {
-    pub signatures: Vec<Signature>,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct Data {
-    pub txs: Option<Vec<String>>,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct Block {
-    pub header: BlockHeader,
-    pub data: Data,
-    pub last_commit: LastCommit,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct BlockRPC {
-    pub block_id: BlockId,
-    pub block: Block,
-}
+pub struct BlockRPC {}
 
 impl BlockRPC {
-    pub async fn load_height(url: String, height: i64) -> Result<Self> {
+    pub async fn load_height(url: String, height: i64) -> Result<module_block_rpc> {
         let url = format!("{}/block?height={}", url, height);
 
         let r = reqwest::get(url)
             .await?
-            .json::<RPCResponse<BlockRPC>>()
+            .json::<RPCResponse<module_block_rpc>>()
             .await?;
         Ok(r.result)
     }
