+++ scanner/src/utils/validator.rs
@@ -1,35 +1,18 @@
-use serde::Deserialize;
-
 use crate::Result;
 
 use super::RPCResponse;
 
-#[derive(Deserialize, Debug)]
-pub struct PubKey {
-    pub r#type: String,
-    pub value: String,
-}
+use module::rpc::validator::ValidatorsRPC as module_validators_rpc;
 
-#[derive(Deserialize, Debug)]
-pub struct Validator {
-    pub address: String,
-    pub pub_key: PubKey,
-    pub voting_power: String,
-    pub proposer_priority: String,
-}
-
-#[derive(Deserialize, Debug)]
-pub struct ValidatorsRPC {
-    pub validators: Vec<Validator>,
-}
+pub struct ValidatorsRPC {}
 
 impl ValidatorsRPC {
-    pub async fn load_height(url: String, height: i64) -> Result<Self> {
+    pub async fn load_height(url: String, height: i64) -> Result<module_validators_rpc> {
         let url = format!("{}/validators?height={}", url, height);
 
         let r = reqwest::get(url)
             .await?
-            .json::<RPCResponse<ValidatorsRPC>>()
+            .json::<RPCResponse<module_validators_rpc>>()
             .await?;
         Ok(r.result)
     }
